Extract dropout filter.
There are now two bikes that have spurious zero issues. Flywheel has the
issue with power. Keiser has the issue with both power and cadence. The
same filter can be used for both bikes and remove some duplication in
the code. Also adds some tests. Fixes #56.
